+++ war/src/main/webapp/WEB-INF/var/db/sql/schema/mysql/jahia-schema.sql
@@ -22,7 +22,7 @@
     drop table if exists jahia_version;
 
     create table jahia_contenthistory (
-        id bigint not null,
+        id integer not null,
         entry_action varchar(255),
         entry_date datetime,
         message varchar(255),
