+++ war/src/main/webapp/WEB-INF/var/db/sql/schema/postgresql/jahia-schema.sql
@@ -20,7 +20,7 @@
     drop table jahia_version;
 
     create table jahia_contenthistory (
-        id int8 not null,
+        id int4 not null,
         entry_action varchar(255),
         entry_date timestamp,
         message varchar(255),
